stories

+++ 1.-FRONT/goodvibing_app/src/DeleteStory.js
@@ -1,29 +1,37 @@
-import { useSelector } from "react-redux";
-import { NavLink, useParams } from "react-router-dom";
+import { useSelector, useDispatch } from "react-redux";
+import { useParams, useHistory } from "react-router-dom";
 
 function DeleteStory() {
   const token = useSelector((s) => s.user?.token);
   const { id } = useParams();
+  const history = useHistory();
+  const dispatch = useDispatch();
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    const res = await fetch(`http://localhost:4000/api/stories/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: "Bearer " + token,
-      },
-    });
-    if (res.ok) {
+    if (window.confirm("¿Estás seguro de eliminar tu historia?")) {
+      const res = await fetch(`http://localhost:4000/api/stories/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: "Bearer " + token,
+        },
+      });
       const data = await res.json();
-      console.log({ data });
+
+      if (res.ok) {
+        history.push("/mystories");
+      } else {
+        dispatch({ type: "SET_ERROR", message: data.message });
+      }
     }
   };
 
   return (
-    <div className="deletestory" onSubmit={handleSubmit}>
-      <NavLink to="/mystories">Eliminar historia</NavLink>
-    </div>
+    <form onSubmit={handleSubmit}>
+      <h1>Borrar tu historia</h1>
+      <button>Eliminar historia</button>
+    </form>
   );
 }
